db 600042
* Add user preference `bookmark_latest`, Only keep the latest media bookmark

+++ src/Repository/Model/Bookmark.php
@@ -24,6 +24,7 @@ declare(strict_types=0);
 
 namespace Ampache\Repository\Model;
 
+use Ampache\Config\AmpConfig;
 use Ampache\Module\System\Core;
 use Ampache\Module\System\Dba;
 use PDOStatement;
@@ -92,11 +93,11 @@ class Bookmark extends database_object
     }
 
     /**
-     * get_bookmark
+     * get_bookmarks
      * @param array $data
      * @return integer[]
      */
-    public static function get_bookmark($data)
+    public static function getBookmarks($data)
     {
         $bookmarks   = array();
         if ($data['object_type'] !== 'bookmark') {
@@ -125,9 +126,11 @@ class Bookmark extends database_object
     public static function create(array $data, int $userId, int $updateDate)
     {
         $comment = scrub_in($data['comment']);
-        // delete duplicates first
-        $sql = "DELETE FROM `bookmark` WHERE `user` = ? AND `comment` = ? AND `object_type` = ? AND `object_id` = ?;";
-        Dba::write($sql, array($userId, $comment, $data['object_type'], $data['object_id']));
+        if (AmpConfig::get('bookmark_latest', false)) {
+            // delete duplicates first
+            $sql = "DELETE FROM `bookmark` WHERE `user` = ? AND `comment` = ? AND `object_type` = ? AND `object_id` = ?;";
+            Dba::write($sql, array($userId, $comment, $data['object_type'], $data['object_id']));
+        }
 
         //insert the new bookmark
         $sql = "INSERT INTO `bookmark` (`user`, `position`, `comment`, `object_type`, `object_id`, `creation_date`, `update_date`) VALUES (?, ?, ?, ?, ?, ?, ?)";
